Patch up rpncomm to use a non-assembler version of the timing routine.
Code adapted from a similar (I hope!) routine in perftools.

Also, freeze the rpncomm version on a version that is known to work with this
recipe.  Some recent commits (Jan 2019) seem to be causing some mismatch in
the expected rpncomm library name, causing a build failure.  I still need to
investigate the source of this problem (could be an issue with the recipe, not
rpncomm).